requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-cors-headers==3.5.0
django-filter==2.4.0
django-filters==0.2.1
django-rest-swagger==2.2.0
django-simple-history==2.12.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
lml==0.1.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
openpyxl==3.0.5
Pillow==8.0.1
pyexcel-io==0.6.4
pyexcel-xlsx==0.6.0
PyJWT==1.7.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
